Removed an unused interface and changed the constructors for position to double

+++ oop15-johnny2d/src/it/unibo/oop/model/Wall.java
@@ -4,7 +4,7 @@ import static it.unibo.oop.utilities.CharactersSettings.*;
 
 public class Wall extends AbstractEntity{
 	
-	public Wall(int startingX, int startinY) {
+	public Wall(double startingX, double startinY) {
 		super(startingX, startinY);
 	}
